+++ Simulation.h
@@ -28,7 +28,7 @@ class Simulation {
 // =====================================================================
 //                           Public Function members
 // =====================================================================
-	int run();
+	float run();
 	bool isExtinct(); // tests whether the population is extinct (returns 1 if this is tha case)
 	unsigned int state(); // tests whether the state is "extinction"(=0), "exclusion"(=1) or "cohabitation"(=2)
 	void reset(unsigned int height, unsigned int width, float ainit,
